fix: access raw BME280 data
BLE MFG data is not compatible with Ruuvi RAW v2 data packet. Missing from packets is movement counter due to missing interrupt from LIS2DH12.

+++ src/bme280_handler/bme280_handler.h
@@ -7,9 +7,9 @@
 #define BME280_HANDLER_H_
 
 void bme280_fetch(void);
-int16_t bme280_get_temp(void);
+int32_t bme280_get_temp(void);
 uint16_t bme280_get_press(void);
-uint16_t bme280_get_humidity(void);
+uint32_t bme280_get_humidity(void);
 bool init_bme280(void);
 
 #endif /* BME280_HANDLER_H_ */
